Resource API: renaming DefaultResource to Resource
The precedent that exists is to name both the interface and the
implementation in the sdk the same. This avoids the generally
unhelpful "Default" prefix.

opentelemetry-python is standardizing on Google style docstrings. As such
resolving formatting that is inconsistent with the style guilde.

Removing init from the abstract Resource class: it does not need to be
part of the Resource API spec. Other Resource implementations may
consume other arguments as well.