Adding the Resource API (#61)
Design Decisions include:

The API package should be minimal, and example implementations should
live in SDK when possible.

The precedent that exists is to name both the interface and the
implementation in the sdk the same. This avoids the generally
unhelpful "Default" prefix.

opentelemetry-python is standardizing on Google style docstrings. As such
resolving formatting that is inconsistent with the style guilde.