sdk: create_span() and start_span() with span_id
OTEP-7 ("Remove support to report out-of-band telemetry from the API")
states that the API does not support specifying the span_id in
create_span() or start_span(). Also, following OTEP-2 ("Remove
SpanData"), we cannot use the SpanData to specify the span_id.

However, the need to specify the span_id remains for some applications.
Following the suggestion from OTEP-7, I am adding this feature in the
SDK only.

This patch adds the named argument span_id in create_span() and
start_span() to optionally specify the id of the span to be created.
When unset, the previous behavior remains and the SDK allocates a
span_id automatically.

Callers should give the name of the argument and not rely on the
current argument order because more arguments could be passed in the
future. OTEP-2 lists possible additional optional arguments:
- start timestamp
- span ID
- resource

Example of call:
```
    with tracer.start_span('foo', span_id=my_span_id):
        print(Context)
```

OTEP-7: https://github.com/open-telemetry/oteps/pull/26
OTEP-2: https://github.com/open-telemetry/oteps/blob/master/text/0002-remove-spandata.md