Addressing feedback
Migrating formatters outside of context. The design goal is to
allow context to be it's own object agnostic of the requirements
of specific telemetry. Instead, moving it under a broader propagators
module.

Removing UnifiedContext in lieu of directly using Context.
UnifiedContext adds an unneeded layer of abstraction that would
then require more operations to migrate into context.

Removing binaryformat and httptextformat modules specific to
SpanContext and DistributedContext. These are now handled by
a single formatter.