sdk: add Exporter interface, SimpleSpanProcessor and InMemorySpanExporter (#119)
* sdk/trace: add Exporter interface and SimpleExportSpanProcessor

Exporter is an interface that allows different services to export recorded
spans in its own format. SimpleExportSpanProcessor is an implementation of
SpanProcessor that passes ended spans directly to a configured Exporter.

The current interface for exporters directly receives an SDK Span, it could
be improved in the future to receive a different object containing a
representation of the Span.

* sdk/trace/exporter: add InMemorySpanExporter

InMemorySpanExporter is a simple implementation of the Exporter interface
that saves the exported spans in a list in memory.  This class is useful
for testing purposes.