sdk/trace/span: comply with concurrency specification (#131)
* sdk/trace/span: comply with concurrency specification

According to the spec [1], all methods in span must be thread safe.  This
commit guarantees that by adding a lock in the places it was missing.

[1] https://github.com/open-telemetry/opentelemetry-specification/blob/d34318277b1999edccf5ab1d63024ec172ae5e49/specification/concurrency.md

* sdk/trace/span: ignore call to methods after span is ended

Calls to methods like set_attribute, add_event, add_link and so on should not
happen after the span has been ended.  This commit introduces some checks to
ignore and print a warning when those calls happen.

* avoid holding a lock while logging