sdk/trace/exporters: add batch span processor exporter
The exporters specification states that two built-in span processors should be
implemented, the simple processor span and the batch processor span.

This commit implements the later, it is mainly based on the opentelemetry/java
one.

The algorithm implements the following logic:
- a condition variable is used to notify the worker thread in case the queue
is half full, so that exporting can start before the queue gets full and spans
are dropped.
- export is called each schedule_delay_millis if there is a least one new span
to export.
- when the processor is shutdown all remaining spans are exported.