ext/jaeger: fix bad variable type
The type of startTime and duration were float types, it caused an exception
in the thrift package because they are declared as i64 integers in thrift.
It wasn't detected because there were not tests covering it, also for
some reason I have been using an old version on my computer all this time.

This commit fixes it by ensuring the datatypes are int (changing 1eX notation
by 10 ** X notation), it also intruce new tests to guarantee that a similar
problem is detected the next time.