Don't show inheritance in docs
Sphinx is trying to resolve references such as `opentracing.Span`
when `show-inheritance` is enabled. However, the actual modules live
in places like `opentracing.span.Span` and are *imported* in
`__init__.py` at the root of the `opentracing` package. This makes
Sphinx try to link to `opentracing.span.Span` instead of
`opentracing.Span` which fails because the OpenTracing docs don't
contain documentation for `opentracing.span.Span` since most likely
aren't expected to use that path in imports.

This is a workaround until we find a better solution.