sdk: shut down span processors automatically
The BatchExportSpanProcessor is an asynchronous span processor that uses a
worker thread to call the different exporters.  Before this commit applications
had to shut down the span processor explicitely to guarantee that all the spans
were summited to the exporters, this was not very intuitive for the users.

This commit removes that limitation by implementing the tracer's __del__
method and an atexit hook.  According to __del__'s documentation [1] it is
possible that sometimes it's not called, for that reason the atexit hook is
also used to guarantee that the processor is shut down in all the cases.

[1] https://docs.python.org/3/reference/datamodel.html#object.__del__