Fix black ignoring files on travis.
First, black already parsers .gitignore, so no need for the exclude
option in pyproject.toml (apart from checked in generated files). The
problem with this is that .gitignore has `build/` in it already and
black does not take into account the root of the git repository, but
instead seems to apply the ignore patterns to the path that results from
the command line arguments (be that relative or absolute).