All named tracers now share the same context
The conversation in https://github.com/open-telemetry/opentelemetry-specification/issues/455
is specifying that the tracer is no longer responsible for handling the setting
and getting of active spans. As such, named tracers would only be respnosible
for creating new spans, but not for setting the active one.

This implies that there tracers do not have their own active spans.

In addition, there is a benefit of having a single active span, as it
vastly simplifies the process to modify the current span (no need to
explicitly retrieve the tracer responsible for getting the span).