Adding trace.get_tracer convenience function
It's fairly common to need to retrieve tracers, and as such adding the
additional tracer_source() call to every retrieval of a tracer can add to
a lot of extra boilerplate at minimal semantic value. It would be uncommon
for one to use multiple trace_source objects, as typically one would want
all tracers to be created and behave in a similar way (e.g. passed to the
same span processor).