Moving contextvars and threadlocal context implementations to the API (#419)
While keeping the flexibility of the interface to support additional implementations of the RuntimeContext, moving the ContextVarsRuntimeContext and ThreadLocalRuntimeContext into the API ensures a useful context implementation for users without requiring additional configuration. This changes checks the version number to determine which implementation to use, unless it is overridden by the OPENTELEMETRY_CONTEXT environment variable. Removing the DefaultRuntimeContext as part of this change.

Signed-off-by: Alex Boten <aboten@lightstep.com>
Co-authored-by: Chris Kleinknecht <libc@google.com>