docs: move ext readmes to docstring
This commit moves the text on the readmes of the external packages to their
source code as a docstring. This improves the generated documentation and helps
to consolidate the documentation in a single place.

This commit uniforms all the readmes present on each packager, they only contain
a 1 line description, installation instructions and a link to the online
documentation of such package. This small readme will be the one shown on Github
and on PyPI.

The full documentation will be generated by autodoc and stored in readthedocs
or whatever site we choose.