docs: fix link to stable versions
The "scm_raw_web" and "scm_web" macros are used in the documentation to create
links to specific versions of the documents. It's useful to point to the right
version of the examples for instance.

There is a special "stable" version in readthedocs that points to the highest
release, generated links to that version are wrongly pointing to a non existing
"satble" release. This commit fixes that logic by getting the highest release
by using the Github API when building documents for the stable release.