metrics: Implement release for handles and observers (#435)
This commit implements a solution for releasing instrument handles and
observers.

For the handles it is based on a ref count that is increased each time the
handled is acquired, when the ref count reaches 0 the handle is removed on
collection time.  The direct call convention is updated to release the handle
after it has been updated.

The observer instrument is only updated on collection time, so it can be
removed as soon as the user request to do so.