ext/jaeger: fix exporting to collector
The exporting of traces to the collector is broken, it replies with the
"Unable to process request body: Required field Process is not set" error.

The current implementation is based on OpenCensus [1], what appears to be broken
too, it's not totally clear at this time what's wrong with that.

This commit changes the exporting logic to be similar to the opentelemetry-go [2]
one that is working. The main change is to perform the request directly without
using the client provided by the generated files.

[1] https://github.com/census-instrumentation/opencensus-python/tree/master/contrib/opencensus-ext-jaeger
[2] https://github.com/open-telemetry/opentelemetry-go/blob/master/exporters/trace/jaeger/jaeger.go