ext/jaeger: fix exporting to collector
The exporting of traces to the collector is broken, it replies with the
"Unable to process request body: Required field Process is not set" error.

The current implementation is based on OpenCensus [1], what appears to be broken
too, it's not totally clear at this time what's wrong with that.

This commit changes the logic to avoid using jaeger.Client that appears to be
the broken piece, it is changed by a direct invokation to the THTTPClient
transport from thrift.