sdk: Add support for lazy events and links (#474)
The computation of attributes is expensive in some conditions, even in
some cases those attributes are not used at all. This commit fixes the
add_lazy_event method by providing a true solution that delays the attributes
calculation until attributes are accessed. It also provides a new LazyLink class.

Events are also moved to SDK, as they are not a required object to satisfy the API specification.

Co-authored-by: Leighton Chen <lechen@microsoft.com>