ext: Expect tracer provider instead of tracer in integrations (#602)
Standardize the interface that trace providers are specified in integrations, as specified in #585.

Adding a helper to create and return a configured TracerProvider with a the span
processor and the memory exporter

api: Add tracer provider parameter to trace.get_tracer(). This eliminates the need for a helper function and boilerplate code to retrieve the appropriate tracer from a passed tracer_provider.