api: Add reset for configuration testing (#636)
It is a common occurrence in tests that the global Configuration object needs to be "reset" between tests. This means that its attributes need to be set back to their original values. Since the Configuration object is immutable by design, some additional, non-production available mechanism is needed to perform this action.

The need for this feature was mentioned in a conversation in #630.