declaring opentelemetry-api/tests as a namespace module.
naming the tests module as a namespace package ensures that
relative imports will resolve properly for other test packages,
as it enables searching for a composite of multiple test modules.

only the opentelemetry-api directory needs this code, as it is
the first tests module found by pylint during eachdist.py lint
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.