instrument: moving auto-instrumentation to a submodule
The opentelemetry.instrumentation namespace will be used by
several packages as they become compliant to the opentelemetry specification.

As a result, top-level instrumentations who have a name identical to their
instrumented library (e.g. opentelemetry.instrumentation.requests or
opentelemetry.instrumentation.starlette) will have a cyclic import
error due to opentelemetry-instrument's injection of the
opentelemetry.instrument folder in the path.

Moving auto-instrumentation to a submodule solves this problem, as
the only files in the opentelemetry.instrumentation.auto_instrument
folder is the sitecustomize and the __init__.py itself.