docs: Add propagators section to getting started
It is common for those who are using different propagation protocols
(e.g. B3) to be confused about why their applications are not using
their HTTPTextFormat of choice.

Adding this missing documentation to the getting-started also
indicates to those new to OpenTelemetry that they will need to make
a conscious choice of propagation protocol.