Go back a bit and use the instrumentor
I really dislike this pattern, changing other libraries, but it's the
way that most of these things work, so I'll stick with it.  One can
still use the Interceptor directly if desired.

Also reworked the wrapped server so it's no longer a wrapper, but simply
adds on our interceptor.  This allows adding any custom interceptors as
well.

A caveat here is that if you need to specifically order interceptors,
you'll need to do it manually rather than using the instrumentor-enable
thing.