Make event attributes immutable
This PR makes it more difficult for someone to edit the attributes
of an `Event` once it has been added to a `Span`, as well as
adding some regression tests around the other properties of an
`Event` to ensure they have similar protections.

It isn't particularly straightforward or idiomatic to make objects
truly immutable in Python, but if these steps don't go far enough
let me know!

Fixes #1038