Update bootstrap cmd to use exact version when installing instrumentation packages
This commit updates the bootstrap command to ensure it'll install
instrumentation packages compatible with the opentelemetry-instrument
package.

Since opentelemetry-instrumentation (as as a result all
instrumentations) are not stable yet, installing an instrumentation
package older or newer than opentelemtry-instrumentation can result in
a broken setup. This commit updates the bootstrap command so that it'll
always install instrumentation packages with the exact same version as
it's own.