Fixed B3 Propagator
B3 propagator was returning `None` instead of invalid context after last
change. This resulted in many apps crashing when B3 failed to extract a
valid context from carrier. This PR patches it to not make apps crash
and return an invalid context when one cannot be extracted from the
carrier.