Allow span limits to be unset
Ths spec recommends most span limits to default to 128 and allow users
to change it via environment variables. It does not dictate that limits
should always be set though. So users should be able to remove these
limits. This commit accepts `"none"` as a valid value for the limit
config environment variables.