`opentelemetry-exporter-prometheus`: restore package
The Prometheus exporter was moved into the `metrics` branch prior to releasing 1.0 for tracing. Now that the metrics API and SDK are further along, we can bring back the exporter and update it to use the new methods. The original code for this PR comes from https://github.com/open-telemetry/opentelemetry-python/tree/metrics/exporter/opentelemetry-exporter-prometheus