Add setuptools runtime requirement
The `pkg_resources` module is provided by `setuptools`.
The `setuptools` package is not part of the Python stdlib but is often
available in the system environment as a build-time requirement.

Explicitly listing `setuptools` as a requirement protects
`opentelemetry-python` from breaking with import errors in cases where
`setuptools` is not available system-wide.

This commit pins the version to `setuptools >= 16.0` which is the first
release that includes the required imports (2015).