Detect and backoff version in metrics exporter
Since backoff==2.0.0, the API of the expo function has changed. The
"porcelain" methods of the backoff library (the decorators
backoff.on_exception and backoff.on_predicate) send a "None" value into
the generator and discard the first yielded value. This is for
compatibility with the more general wait generator API inside the
backoff package.

This commit allows the HTTP OTLP metrics exporter to automatically
detect the behavior of the installed backoff package and adapt
accordingly.