PeriodicExportingMetricReader will continue collection times out (#3098)
In cases where collection times out, the period exporting reader thread should
not terminate, but instead catch, log, and continue on after the regular interval
seconds.

Prior to this commit, a metric collection timeout would terminate the thread and
stop reporting metrics to the wrapped exporter resulting in the appearance in
observability tooling of metrics just stopping without reason.