Make BoundedAttributes constructor more efficient
There is no risk of locking issues during a constructor call. It is also
not desirable to lock-unlock-... for each provided attribute to the
constructor.

This commit seeks to improve on these by refactoring __setitem__ such
that the 'body' inside the lock context manager is moved to a separate
method _unsafe__setitem. This method can then be called by internal
methods without taking the lock first.

Another method 'update' (overridden from MutableMapping) is added that
takes the lock *once* and then calls another new method _unsafe_update,
which in turn simply iterates the attribute items and calls
_unsafe__setitem on them.