Add Warning for Endpoint Mismatch in OTLP GRPC Exporter
This commit introduces a warning message in the OTLP GRPC Exporter when there is a mismatch between the configured endpoint and the provided endpoint. This change aims to prevent silent truncation of the endpoint value, which could lead to confusion and time-consuming debugging for users trying to configure a HTTP exporter but mistakenly passing a GRPC one. This issue might be a common occurrence for newcomers, as almost all the examples use the GRPC exporter. The warning message provides clear, actionable feedback to the user, improving the overall user experience.

Addresses https://github.com/open-telemetry/opentelemetry-python/issues/3619