[DO NOT MERGE] Port of jaeger remote sampler support
- Introduce a dependency on tornado, in order to use IOLoop
  in RemoteControlledSampler
- Introduce a variety of new samplers, which support the
  jaeger remote sampling protocol

DRAFT - DO NOT MERGE

There are a variety of challenging TODO's and, no doubt,
changes to be made based on feedback from project maintainers.

The goal of this initial set of changes is to conduct a pretty
"straight" port of the corresponding code in jaeger-client-python,
to get discussion going on how opentelemetry-python support for
jaeger remote sampling should ultimately work.