Add test cases for RetryableExporter
A few tests are still commented out and will be enabled later, once
the implementation is updated to implement the desired behavior.

The tests are, like other existing tests, timing sensitive which makes
them prone to fail on slow systems. It would certainly be better to get
rid of this issue. I haven't found another approach in the existing test
cases so I sticked to the current pattern.