GRPC log and span exporter: Use timeout passed to export
The timeout passed as argument to the export method is now overriding
the default values set during initialization if given.

This change is made to align the API of all exporters. In the metric
exporter this parameter was already existant but was not used. This was
changed with the previous commit (2102e3b25b52).