Align HTTP metric exporter timeout with other exporters
It's desirable to have a common API across all exporters.

Since the previously existing timeout_millis argument wasn't used
anyways, it should not be an issue to change the default value here.

The None case will fall back to the object level timeout, which is
the same as the previous default argument value.