api: add options for a bit more liberal parsing in parse_env_headers
Add a liberal boolean parameter to parse_env_headers default to False to
keep backward compatibility for the function.
If liberal=True we try to parse the string even if it does comply with
the the W3C Baggage HTTP header format to be more compatible with other
other languages SDKs that accept non URL-encoded headers by default.