Change export lock to cover the whole export procedure
Previously the export lock was released by the export between retries,
which gave a shutdown the opportunity to interrupt the export depending
on which thread is first to get the lock again. This behavior is not
desired as it heavily relies on timing and makes things less
deterministic.

The retry loop still checks whether the shutdown flag is set and will
break in case it was told to do so.