Enable exporter shutdown timeout to abort ongoing export
Replace the RetryableExporter shutdown flag and sleep with a
`threading.Event` to make the waiting time for the next retry
interruptable.

This enables us to change the semantic of the shutdown timeout
to work in a less surprising way. Previously the timeout worked
roughly like this:

1. Wait for the current export iteration to finish
2. Try to get the lock faster than the export thread
3. If the lock was acquired, the shutdown flag was set
4. If the lock was not acquired, the shutdown flag was set
5. The export thread would have picked this flag up after finishing the
   current retry cycle.

This lead to hanging applications on shutdown, giving the applications
no choice to influence the shutdown behavior. Uses could not decide how
much time shall be spent trying to prevent the loss of data.

This change allows shutdown to interrupt the current export waiting call
so that the exporter can be shut down with a user provided deadline.