Change set_attributes to take a Mapping
This currently specifies a Dict[str, types.AttributeValue], which is invariant in its value type. As a result, you cannot pass a dictionary with a narrower value type, e.g. a dict[str, str]. But this should be allowed, since set_attributes simply adds all key-value pairs in its argument, and all such pairs are valied. Using Mapping takes care of this issue, since it's covariant in its value type.