Delete duplicated OTLP Exporter tests, move them to the mixin unit test. Fix broken shutdown unit test. (#4504)
* Move all OTLP exporter tests that are testing the
underlying behavior in the mixin to the mixin unit
tests, instead of having them specified multiple
times in the Metric/Log/Trace exporters.

Fix the shutdown tests which were flaky, so that
they just test whether a pending export call completes
or not.  Update shutdown so it doesn't release the
lock -- in cases where an export call is pending,
export then also releases the lock causing a Runtime
Error: https://docs.python.org/3/library/threading.html#threading.Lock.release.

* Use threading.Event() to communicate when an
export RPC is occuring, so that shutdown waits for
export RPC to finish.

Use threading.Event() to communicate when shutdown
is occuring, so that sleep is interrupted if a
shutdown is occuring.

* Remove changes to shutdown logic. Will do a
separate PR for those.

* Address comments in PR

* Fix lint issue

* Address more comments

* Respond to review comments

* Respond to comments on PR

* Remove shutdown call from test class, so it just
falls through to mixin.

* Some cleanup

* Respond to comments

* Fix precommit

* Add back missing test